Refactors package structure for consent-related classes and updates imports
