+++ accountservice/src/main/java/com/micronauticals/accountservice/repository/ConsentRepository.java
@@ -1,6 +1,6 @@
 package com.micronauticals.accountservice.repository;
 
-import com.micronauticals.accountservice.entity.Consent;
+import com.micronauticals.accountservice.entity.consent.Consent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ConsentRepository extends JpaRepository<Consent, String> {
